+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -  `keyword.operator.practice` for operators
     -  `string.other.path.practice` for unquoted paths
     -  `comment.practice` for comments
+    -  `constant.numeric.practice` for numbers
 
 ### Changed